roachprod: fix behavior when mixing GCE projects
Currently roachprod has very poor behavior when used with different
projects on the same host. For example:
```
shell1: GCE_PROJECT=andrei-jepsen roachstress.sh ... // this will run ~forever
sometime later in shell2: roachprod sync (on the default project)
```

The sync on the default project removes the cached information for the
cluster on `andrei-jepsen`, which causes `roachprod` commands against
that cluster (from within the `roachstress.sh` script) to fail.

We fix this by ignoring any cached clusters that reference a project
that the provider was not configured for - both when loading clusters
into memory and when deleting stale cluster files during `sync`.

As part of the change, we also improve the output of `list` to remove
the colon after the cluster name and to include the GCE project:

```
$ roachprod list --gce-project cockroach-ephemeral,andrei-jepsen
Syncing...
Refreshing DNS entries...
glenn-anarest                  [aws]                      9  (142h41m39s)
glenn-drive                    [aws]                      1  (141h41m39s)
jane-1635868819-01-n1cpu4      [gce:cockroach-ephemeral]  1  (10h41m39s)
lin-ana                        [aws]                      9  (178h41m39s)
local-foo                      [local]                    4  (-)
radu-foo                       [gce:andrei-jepsen]        4  (12h41m39s)
radu-test                      [gce:cockroach-ephemeral]  4  (12h41m39s)
```

Release note: None

+++ pkg/roachprod/clusters_cache.go
@@ -21,6 +21,7 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/roachprod/cloud"
 	"github.com/cockroachdb/cockroach/pkg/roachprod/config"
 	"github.com/cockroachdb/cockroach/pkg/roachprod/install"
+	"github.com/cockroachdb/cockroach/pkg/roachprod/vm"
 	"github.com/cockroachdb/cockroach/pkg/roachprod/vm/local"
 	"github.com/cockroachdb/cockroach/pkg/util/log"
 	"github.com/cockroachdb/errors"
@@ -96,6 +97,19 @@ func loadCluster(name string) (*cloud.Cluster, error) {
 	return c, nil
 }
 
+// shouldIgnoreCluster returns true if the cluster references a project that is
+// not active. This is relevant if we have a cluster that was cached when
+// another project was in use.
+func shouldIgnoreCluster(c *cloud.Cluster) bool {
+	for i := range c.VMs {
+		provider, ok := vm.Providers[c.VMs[i].Provider]
+		if !ok || !provider.ProjectActive(c.VMs[i].Project) {
+			return true
+		}
+	}
+	return false
+}
+
 // LoadClusters reads the cached cluster metadata from config.ClustersDir and
 // populates install.Clusters.
 func LoadClusters() error {
@@ -114,6 +128,9 @@ func LoadClusters() error {
 
 		if len(c.VMs) == 0 {
 			return errors.Errorf("found no VMs in %s", clusterFilename(name))
+		}
+		if shouldIgnoreCluster(c) {
+			continue
 		}
 
 		sc := &install.SyncedCluster{
@@ -155,9 +172,17 @@ func syncClustersCache(cloud *cloud.Cloud) error {
 	}
 	for _, name := range clusterNames {
 		if _, ok := cloud.Clusters[name]; !ok {
-			filename := clusterFilename(name)
-			if err := os.Remove(filename); err != nil {
-				log.Infof(context.Background(), "failed to remove file %s", filename)
+			// This cluster may no longer exist, or it may involve projects that are
+			// not active in the current invocation.
+			c, err := loadCluster(name)
+			if err != nil {
+				return err
+			}
+			if !shouldIgnoreCluster(c) {
+				filename := clusterFilename(name)
+				if err := os.Remove(filename); err != nil {
+					log.Infof(context.Background(), "failed to remove file %s", filename)
+				}
 			}
 		}
 	}
